os/,osd/: restructure the rados name length check
Rather than just checking the length of the name, check the sum of the
name, the locator, and the namespace.  Also, rather than ask the
ObjectStore implementation, pass the head object to the implementation
to validate.  This allows LFNIndex to account for the idiosyncracies of
its filename escaping and for different xattr value max sizes.

Signed-off-by: Samuel Just <sjust@redhat.com>